移除一些几乎用不着的api
