+++ Masuit.Tools.Core/Masuit.Tools.Core.csproj
@@ -30,9 +30,6 @@
 
   <ItemGroup>
     <Compile Include="..\Masuit.Tools.Abstractions\Config\CoreConfig.cs" Link="Config\CoreConfig.cs" />
-    <Compile Include="..\Masuit.Tools.Abstractions\Database\DataExt.cs">
-      <Link>Database\DataExt.cs</Link>
-    </Compile>
     <Compile Include="..\Masuit.Tools.Abstractions\Database\DataTableBuilder.cs">
       <Link>Database\DataTableBuilder.cs</Link>
     </Compile>
@@ -47,12 +44,6 @@
     </Compile>
     <Compile Include="..\Masuit.Tools.Abstractions\DateTimeExt\DateTimeHelper.cs">
       <Link>DateTimeExt\DateTimeHelper.cs</Link>
-    </Compile>
-    <Compile Include="..\Masuit.Tools.Abstractions\DateTimeExt\DateUtil.cs">
-      <Link>DateTimeExt\DateUtil.cs</Link>
-    </Compile>
-    <Compile Include="..\Masuit.Tools.Abstractions\DateTimeExt\TimeHelper.cs">
-      <Link>DateTimeExt\TimeHelper.cs</Link>
     </Compile>
     <Compile Include="..\Masuit.Tools.Abstractions\DateTimeExt\WeekHolidayStruct.cs">
       <Link>DateTimeExt\WeekHolidayStruct.cs</Link>
@@ -104,9 +95,6 @@
     </Compile>
     <Compile Include="..\Masuit.Tools.Abstractions\Files\INIFile.cs">
       <Link>Files\INIFile.cs</Link>
-    </Compile>
-    <Compile Include="..\Masuit.Tools.Abstractions\Files\WinrarHelper.cs">
-      <Link>Files\WinrarHelper.cs</Link>
     </Compile>
     <Compile Include="..\Masuit.Tools.Abstractions\Hardware\CpuInfo.cs">
       <Link>Hardware\CpuInfo.cs</Link>
